backend/requirements/base.txt
fastapi[all]
peewee
python-dotenv
psycopg2-binary